move ProgressDialog.js to src/gui/dialogs/
