+++ src/calendar/view/CalendarEventEditDialog.js
@@ -39,7 +39,7 @@ import {showBusinessFeatureRequiredDialog} from "../../misc/SubscriptionDialogs"
 import {locator} from "../../api/main/MainLocator"
 import {BusinessFeatureRequiredError} from "../../api/main/BusinessFeatureRequiredError"
 import type {MailboxDetail} from "../../mail/model/MailModel"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {PasswordIndicator} from "../../gui/PasswordIndicator"
 import {TimePicker} from "../../gui/TimePicker"
 import type {ContactModel} from "../../contacts/model/ContactModel"
