+++ src/contacts/view/ContactView.js
@@ -27,7 +27,7 @@ import {vCardFileToVCards, vCardListToContacts} from "../VCardImporter"
 import {LockedError, NotFoundError} from "../../api/common/error/RestError"
 import {MultiContactViewer} from "./MultiContactViewer"
 import {BootIcons} from "../../gui/base/icons/BootIcons"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {locator} from "../../api/main/MainLocator"
 import {ContactMergeView} from "./ContactMergeView"
 import {getMergeableContacts, mergeContacts} from "../ContactMergeUtils"
