+++ src/file/FileController.js
@@ -4,7 +4,7 @@ import {worker} from "../api/main/WorkerClient"
 import {convertToDataFile, createDataFile} from "../api/common/DataFile"
 import {assertMainOrNode, isAndroidApp, isApp, isDesktop} from "../api/common/Env"
 import {downcast, neverNull} from "../api/common/utils/Utils"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {CryptoError} from "../api/common/error/CryptoError"
 import {lang} from "../misc/LanguageViewModel"
 import {BrowserType} from "../misc/ClientConstants"
