+++ src/login/ExternalLoginView.js
@@ -20,7 +20,7 @@ import {lang} from "../misc/LanguageViewModel"
 import {keyManager} from "../misc/KeyManager"
 import {client} from "../misc/ClientDetector"
 import {windowFacade} from "../misc/WindowFacade"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {CloseEventBusOption, Keys} from "../api/common/TutanotaConstants"
 import {progressIcon} from "../gui/base/Icon"
 import {ButtonN, ButtonType} from "../gui/base/ButtonN"
