+++ src/login/LoginViewController.js
@@ -25,7 +25,7 @@ import type {LoginView} from "./LoginView"
 import {deviceConfig} from "../misc/DeviceConfig"
 import {client} from "../misc/ClientDetector"
 import {secondFactorHandler} from "../misc/SecondFactorHandler"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {CancelledError} from "../api/common/error/CancelledError"
 import {notifications} from "../gui/Notifications"
 import {isMailAddress} from "../misc/FormatValidator"
