+++ src/login/recover/RecoverLoginDialog.js
@@ -5,7 +5,7 @@ import type {ButtonAttrs} from "../../gui/base/ButtonN"
 import {ButtonN, ButtonType} from "../../gui/base/ButtonN"
 import {createDropdown} from "../../gui/base/DropdownN.js"
 import {AccessBlockedError, AccessDeactivatedError, NotAuthenticatedError, TooManyRequestsError} from "../../api/common/error/RestError"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {isMailAddress} from "../../misc/FormatValidator"
 import {Type} from "../../gui/base/TextFieldN"
 import {lang} from "../../misc/LanguageViewModel"
