+++ src/mail/editor/MailEditor.js
@@ -40,7 +40,7 @@ import {ExpanderButtonN, ExpanderPanelN} from "../../gui/base/Expander"
 import {newMouseEvent} from "../../gui/HtmlUtils"
 import {windowFacade} from "../../misc/WindowFacade"
 import {UserError} from "../../api/main/UserError"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {htmlSanitizer} from "../../misc/HtmlSanitizer"
 import {DropDownSelectorN} from "../../gui/base/DropDownSelectorN"
 import type {Mail} from "../../api/entities/tutanota/Mail"
