+++ src/mail/view/MailViewer.js
@@ -80,7 +80,7 @@ import {MailHeadersTypeRef} from "../../api/entities/tutanota/MailHeaders"
 import {client} from "../../misc/ClientDetector"
 import type {PosRect} from "../../gui/base/Dropdown"
 import {createAsyncDropDownButton, createDropDownButton, DomRectReadOnlyPolyfilled} from "../../gui/base/Dropdown"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import Badge from "../../gui/base/Badge"
 import {FileOpenError} from "../../api/common/error/FileOpenError"
 import type {DialogHeaderBarAttrs} from "../../gui/base/DialogHeaderBar"
