+++ src/misc/ErrorHandlerImpl.js
@@ -25,7 +25,7 @@ import {OutOfSyncError} from "../api/common/error/OutOfSyncError"
 import stream from "mithril/stream/stream.js"
 import {SecondFactorPendingError} from "../api/common/error/SecondFactorPendingError"
 import {secondFactorHandler} from "./SecondFactorHandler"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {IndexingNotSupportedError} from "../api/common/error/IndexingNotSupportedError"
 import {windowFacade} from "./WindowFacade"
 import * as notificationOverlay from "../gui/base/NotificationOverlay"
