+++ src/settings/AddGroupDialog.js
@@ -8,7 +8,7 @@ import type {ValidationResult} from "./SelectMailAddressForm"
 import {SelectMailAddressForm} from "./SelectMailAddressForm"
 import {worker} from "../api/main/WorkerClient"
 import {getGroupTypeName} from "./GroupViewer"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {logins} from "../api/main/LoginController"
 import type {TranslationKey} from "../misc/LanguageViewModel"
 import {lang} from "../misc/LanguageViewModel"
