+++ src/settings/AddUserDialog.js
@@ -13,7 +13,7 @@ import {CustomerInfoTypeRef} from "../api/entities/sys/CustomerInfo"
 import {addAll} from "../api/common/utils/ArrayUtils"
 import {getCustomMailDomains, neverNull} from "../api/common/utils/Utils"
 import {worker} from "../api/main/WorkerClient"
-import {showProgressDialog, showWorkerProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog, showWorkerProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {PreconditionFailedError} from "../api/common/error/RestError"
 import {showBuyDialog} from "../subscription/BuyDialog"
 import stream from "mithril/stream/stream.js"
