+++ src/settings/ContactFormViewer.js
@@ -15,7 +15,7 @@ import {Dialog} from "../gui/base/Dialog"
 import {neverNull} from "../api/common/utils/Utils"
 import {GroupInfoTypeRef} from "../api/entities/sys/GroupInfo"
 import {getDefaultContactFormLanguage} from "../contacts/ContactFormUtils"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {DatePicker} from "../gui/date/DatePicker"
 import {StatisticLogEntryTypeRef} from "../api/entities/tutanota/StatisticLogEntry"
 import {stringToUtf8Uint8Array, timestampToGeneratedId} from "../api/common/utils/Encoding"
