+++ src/settings/DesktopSettingsViewer.js
@@ -4,7 +4,7 @@ import {assertMainOrNode} from "../api/common/Env"
 import {lang} from "../misc/LanguageViewModel"
 import stream from "mithril/stream/stream.js"
 import {Request} from "../api/common/WorkerProtocol.js"
-import {showProgressDialog} from "../gui/ProgressDialog.js"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog.js"
 import {noOp} from "../api/common/utils/Utils"
 import {Icons} from "../gui/base/icons/Icons"
 import type {TextFieldAttrs} from "../gui/base/TextFieldN"
