+++ src/settings/EditSecondFactorsForm.js
@@ -19,7 +19,7 @@ import {contains} from "../api/common/utils/ArrayUtils"
 import {worker} from "../api/main/WorkerClient"
 import QRCode from "qrcode"
 import {GroupInfoTypeRef} from "../api/entities/sys/GroupInfo"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {copyToClipboard} from "../misc/ClipboardUtils"
 import {htmlSanitizer} from "../misc/HtmlSanitizer"
 import type {ButtonAttrs} from "../gui/base/ButtonN"
