+++ src/settings/GroupViewer.js
@@ -19,7 +19,7 @@ import {GroupMemberTypeRef} from "../api/entities/sys/GroupMember"
 import {logins} from "../api/main/LoginController"
 import {UserTypeRef} from "../api/entities/sys/User"
 import {Icons} from "../gui/base/icons/Icons"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {AdministratedGroupTypeRef} from "../api/entities/sys/AdministratedGroup"
 import {localAdminGroupInfoModel} from "./LocalAdminGroupInfoModel"
 import stream from "mithril/stream/stream.js"
