+++ src/settings/IdentifierListViewer.js
@@ -11,7 +11,7 @@ import {logins} from "../api/main/LoginController"
 import {Icons} from "../gui/base/icons/Icons"
 import {PushServiceType} from "../api/common/TutanotaConstants"
 import {getCleanedMailAddress} from "../misc/Formatter"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {worker} from "../api/main/WorkerClient"
 import {Dialog} from "../gui/base/Dialog"
 import {NotFoundError} from "../api/common/error/RestError"
