+++ src/settings/ImportUsersViewer.js
@@ -4,7 +4,7 @@ import {Dialog} from "../gui/base/Dialog"
 import {lang} from "../misc/LanguageViewModel"
 import {isMailAddress} from "../misc/FormatValidator"
 import {worker} from "../api/main/WorkerClient"
-import {showWorkerProgressDialog} from "../gui/ProgressDialog"
+import {showWorkerProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {BookingItemFeatureType} from "../api/common/TutanotaConstants"
 import {contains} from "../api/common/utils/ArrayUtils"
 import {PreconditionFailedError} from "../api/common/error/RestError"
