+++ src/settings/MailSettingsViewer.js
@@ -16,7 +16,7 @@ import {logins} from "../api/main/LoginController"
 import {getDefaultSenderFromUser, getFolderName} from "../mail/model/MailUtils"
 import {Icons} from "../gui/base/icons/Icons"
 import {worker} from "../api/main/WorkerClient"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import type {MailboxDetail} from "../mail/model/MailModel"
 import {locator} from "../api/main/MainLocator"
 import stream from "mithril/stream/stream.js"
