+++ src/settings/SetDomainCertificateDialog.js
@@ -8,7 +8,7 @@ import {fileController} from "../file/FileController"
 import {utf8Uint8ArrayToString} from "../api/common/utils/Encoding"
 import {InvalidDataError, LockedError, PreconditionFailedError} from "../api/common/error/RestError"
 import {Icons} from "../gui/base/icons/Icons"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {isDomainName} from "../misc/FormatValidator"
 import type {DropDownSelectorAttrs} from "../gui/base/DropDownSelectorN"
 import {DropDownSelectorN} from "../gui/base/DropDownSelectorN"
