+++ src/settings/emaildomain/VerifyOwnershipPage.js
@@ -4,7 +4,7 @@ import m from "mithril"
 import type {AddDomainData} from "./AddDomainWizard"
 import {createDnsRecordTable} from "./AddDomainWizard"
 import {worker} from "../../api/main/WorkerClient"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {lang} from "../../misc/LanguageViewModel"
 import {Dialog} from "../../gui/base/Dialog"
 import {getCustomMailDomains} from "../../api/common/utils/Utils"
