+++ src/settings/whitelabel/WhitelabelSettingsViewer.js
@@ -17,7 +17,7 @@ import {WhitelabelConfigTypeRef} from "../../api/entities/sys/WhitelabelConfig"
 import type {Theme} from "../../gui/theme"
 import {themeManager} from "../../gui/theme"
 import {progressIcon} from "../../gui/base/Icon"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import type {Booking} from "../../api/entities/sys/Booking"
 import {BookingTypeRef} from "../../api/entities/sys/Booking"
 import type {EntityUpdateData} from "../../api/main/EventController"
