+++ src/sharing/view/GroupSharingDialog.js
@@ -15,7 +15,7 @@ import {attachDropdown} from "../../gui/base/DropdownN"
 import type {ButtonAttrs} from "../../gui/base/ButtonN"
 import {ButtonType} from "../../gui/base/ButtonN"
 import {remove} from "../../api/common/utils/ArrayUtils"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import type {ShareCapabilityEnum} from "../../api/common/TutanotaConstants"
 import {ShareCapability} from "../../api/common/TutanotaConstants"
 import {DropDownSelectorN} from "../../gui/base/DropDownSelectorN"
