+++ src/subscription/InvoiceAndPaymentDataPage.js
@@ -9,7 +9,7 @@ import stream from "mithril/stream/stream.js"
 import type {PaymentMethodTypeEnum} from "../api/common/TutanotaConstants"
 import {getClientType, Keys, PaymentDataResultType} from "../api/common/TutanotaConstants"
 import {worker} from "../api/main/WorkerClient"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {getLazyLoadedPayPalUrl} from "./PaymentDataDialog"
 import {logins} from "../api/main/LoginController"
 import {client} from "../misc/ClientDetector"
