+++ src/subscription/PaymentDataDialog.js
@@ -9,7 +9,7 @@ import {PaymentMethodInput} from "./PaymentMethodInput"
 import {updatePaymentData} from "./InvoiceAndPaymentDataPage"
 import {px} from "../gui/size"
 import {formatNameAndAddress} from "../misc/Formatter"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {getClientType, getPaymentMethodType, PaymentMethodType} from "../api/common/TutanotaConstants"
 import {LazyLoaded} from "../api/common/utils/LazyLoaded"
 import {serviceRequest} from "../api/main/Entity"
