+++ src/subscription/PaymentMethodInput.js
@@ -9,7 +9,7 @@ import {PaymentMethodType} from "../api/common/TutanotaConstants"
 import {CreditCardAttrs, CreditCardInput} from "./CreditCardInput"
 import {PayPalLogo} from "../gui/base/icons/Icons"
 import {LazyLoaded} from "../api/common/utils/LazyLoaded"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import type {AccountingInfo} from "../api/entities/sys/AccountingInfo"
 import {AccountingInfoTypeRef} from "../api/entities/sys/AccountingInfo"
 import {locator} from "../api/main/MainLocator"
