+++ src/subscription/SignupForm.js
@@ -17,7 +17,7 @@ import {neverNull} from "../api/common/utils/Utils"
 import {lang} from "../misc/LanguageViewModel"
 import type {DialogHeaderBarAttrs} from "../gui/base/DialogHeaderBar"
 import {DialogHeaderBar} from "../gui/base/DialogHeaderBar"
-import {showWorkerProgressDialog} from "../gui/ProgressDialog"
+import {showWorkerProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {worker} from "../api/main/WorkerClient"
 import {AccessDeactivatedError, AccessExpiredError, InvalidDataError} from "../api/common/error/RestError"
 import {createRegistrationCaptchaServiceGetData} from "../api/entities/sys/RegistrationCaptchaServiceGetData"
