+++ src/subscription/UpgradeConfirmPage.js
@@ -10,7 +10,7 @@ import type {PaidSubscriptionTypeEnum} from "../api/common/TutanotaConstants"
 import {AccountType, Const, PaidSubscriptionType} from "../api/common/TutanotaConstants"
 import {SysService} from "../api/entities/sys/Services"
 import {serviceRequestVoid} from "../api/main/Entity"
-import {showProgressDialog} from "../gui/ProgressDialog"
+import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {worker} from "../api/main/WorkerClient"
 import {HttpMethod} from "../api/common/EntityFunctions"
 import type {UpgradeSubscriptionData} from "./UpgradeSubscriptionWizard"
