+++ src/subscription/giftcards/PurchaseGiftCardDialog.js
@@ -7,7 +7,7 @@ import {serviceRequest} from "../../api/main/Entity"
 import {logins} from "../../api/main/LoginController"
 import type {AccountingInfo} from "../../api/entities/sys/AccountingInfo"
 import {worker} from "../../api/main/WorkerClient"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {GiftCardTypeRef} from "../../api/entities/sys/GiftCard"
 import {locator} from "../../api/main/MainLocator"
 import type {Country} from "../../api/common/CountryList"
