+++ src/subscription/giftcards/RedeemGiftCardWizard.js
@@ -12,7 +12,7 @@ import {LoginForm} from "../../login/LoginForm"
 import type {CredentialsSelectorAttrs} from "../../login/CredentialsSelector"
 import {CredentialsSelector} from "../../login/CredentialsSelector"
 import {deviceConfig} from "../../misc/DeviceConfig"
-import {showProgressDialog} from "../../gui/ProgressDialog"
+import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {worker} from "../../api/main/WorkerClient"
 import {client} from "../../misc/ClientDetector"
 import {ButtonN, ButtonType} from "../../gui/base/ButtonN"
